fix: temporary fix to make extension build
